Add the new ranging accuracty tests to CTS-V-HOST.
Also fix issues when running the latest test code.

Main changes:

* Fix the typo to submit performance value to MPC.
* Modify AndroidManifest.xml to NOT skip this test module when there's
  no UWB module. This is because not all tests require UWB.
* Set device screen to stay on, otherwise BLE RSSI related tests will fail.

Flag: TEST_ONLY
Bug: 399422261
Test: http://sponge2/5adfd0e7-db05-4853-bb16-b5a51974db2e

+++ ranging/tests/multidevices/snippet/uwb/RangingSnippet.java
@@ -18,6 +18,8 @@ package com.google.snippet.ranging;
 
 import android.app.UiAutomation;
 import android.content.Context;
+import android.mediapc.cts.common.PerformanceClassEvaluator;
+import android.mediapc.cts.common.Requirements;
 import android.net.ConnectivityManager;
 import android.net.wifi.WifiManager;
 import android.ranging.MotionState;
@@ -56,8 +58,6 @@ import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Supplier;
-import android.mediapc.cts.common.PerformanceClassEvaluator;
-import android.mediapc.cts.common.Requirements;
 
 public class RangingSnippet implements Snippet {
     private static final String TAG = "GenericRangingSnippet";
@@ -515,13 +515,13 @@ public class RangingSnippet implements Snippet {
     }
 
     @Rpc(description = "Submit MPC BLE Channel Sounding Performance Value")
-    public void submitMpcBleChannelSoundingPerformanceValue(float value) {
+    public void submitMpcBleChannelSoundingPerformanceValue(double value) {
         PerformanceClassEvaluator pce = new PerformanceClassEvaluator(
             "CtsMultiDeviceGenericRangingAccuracyTests:test_channel_sounding_ranging");
         Requirements.Android17BLEChannelSoundingRequirement r7_4_3__h_3_1 =
             Requirements.addR7_4_3__H_3_1().to(pce);
         r7_4_3__h_3_1.setIsBleCsSupported(true);
-        r7_4_3__h_3_1.setBleCsAccuracy(value);
+        r7_4_3__h_3_1.setBleCsAccuracy((float) value);
         pce.submitAndCheck();
     }
